+++ src/docdb/tree/DocDBAccountTreeItemBase.ts
@@ -67,8 +67,14 @@ export abstract class DocDBAccountTreeItemBase extends DocDBTreeItemBase<Databas
     }
 
     private static validateDatabaseName(name: string): string | undefined | null {
-        if (name.length < 1 || name.length > 255) {
+        if (!name || name.length < 1 || name.length > 255) {
             return "Name has to be between 1 and 255 chars long";
+        }
+        if (name.endsWith(" ")) {
+            return "Database name cannot end with space";
+        }
+        if (/[/\\?#]/.test(name)) {
+            return `Database name cannot contain the characters '\\', '/', '#', '?'`;
         }
         return undefined;
     }