+++ src/graph/GraphViewServer.ts
@@ -46,18 +46,11 @@ export class GraphViewServer extends EventEmitter {
   private _httpServer: http.Server;
   private _port: number | undefined;
   private _socket: GraphViewServerSocket;
-  private _previousPageState: {
-    query: string | undefined,
-    results: GraphResults | undefined,
-    errorMessage: string | undefined,
-    view: 'graph' | 'json',
-    isQueryRunning: boolean,
-    runningQueryId: number
-  };
+  private _pageState: PageState;
 
   constructor(private _configuration: GraphConfiguration) {
     super();
-    this._previousPageState = {
+    this._pageState = {
       query: undefined,
       results: undefined,
       errorMessage: undefined,
@@ -131,17 +124,25 @@ export class GraphViewServer extends EventEmitter {
     return Math.max(1, vscode.workspace.getConfiguration().get<number>('cosmosDB.graph.maxEdges'));
   }
 
+  private getViewSettings(): GraphViewSettings {
+    var viewSettings: GraphViewSettings = vscode.workspace.getConfiguration().get<GraphViewSettings>('cosmosDB.graph.viewSettings') || <GraphViewSettings>{};
+    return viewSettings;
+  }
+
   private async queryAndShowResults(queryId: number, gremlinQuery: string): Promise<void> {
     var results: GraphResults | undefined;
     const start = Date.now();
 
     try {
       await callWithTelemetry("cosmosDB.gremlinQuery", async (telemetryProperties, measurements) => {
-        this._previousPageState.query = gremlinQuery;
-        this._previousPageState.results = undefined;
-        this._previousPageState.errorMessage = undefined;
-        this._previousPageState.isQueryRunning = true;
-        this._previousPageState.runningQueryId = queryId;
+        this._pageState = {
+          query: gremlinQuery,
+          results: undefined,
+          errorMessage: undefined,
+          isQueryRunning: true,
+          runningQueryId: queryId,
+          view: this._pageState.view
+        };
 
         measurements.gremlinLength = gremlinQuery.length;
         let stepMatches = gremlinQuery.match("[.]");
@@ -164,7 +165,7 @@ export class GraphViewServer extends EventEmitter {
         };
         measurements.countUniqueVertices = countUniqueVertices;
         measurements.limitedVertices = limitedVertices.length;
-        this._previousPageState.results = results;
+        this._pageState.results = results;
 
         if (results.limitedVertices.length) {
           try {
@@ -185,14 +186,14 @@ export class GraphViewServer extends EventEmitter {
     } catch (error) {
       // If there's an error, send it to the client to display
       var message = this.removeErrorCallStack(error.message || error.toString());
-      this._previousPageState.errorMessage = message;
+      this._pageState.errorMessage = message;
       this._socket.emitToClient("showQueryError", queryId, message);
       return;
     } finally {
-      this._previousPageState.isQueryRunning = false;
+      this._pageState.isQueryRunning = false;
     }
 
-    this._socket.emitToClient("showResults", queryId, results);
+    this._socket.emitToClient("showResults", queryId, results, this.getViewSettings());
   }
 
   private getVertices(queryResults: any[]): GraphVertex[] {
@@ -364,19 +365,19 @@ export class GraphViewServer extends EventEmitter {
   private handleGetPageState(): void {
     this.log('getPageState');
 
-    if (this._previousPageState.query) {
-      this._socket.emitToClient('setPageState', this._previousPageState);
+    if (this._pageState.query) {
+      this._socket.emitToClient('setPageState', this._pageState, this.getViewSettings());
     }
   }
 
   private handleSetQuery(query: string) {
     this.log('setQuery');
-    this._previousPageState.query = query;
+    this._pageState.query = query;
   }
 
   private handleSetView(view: 'graph' | 'json') {
     this.log('setView');
-    this._previousPageState.view = view;
+    this._pageState.view = view;
   }
 
   private handleQueryMessage(queryId: number, gremlin: string) {