+++ src/graph/client/graphTypes.ts
@@ -27,6 +27,37 @@ interface GraphEdge {
 interface GraphVertex {
   id: string;
   type: "vertex";
+  label: string;
+  properties: {
+    [key: string]: [
+      {
+        id: string;
+        value: string;
+      }
+    ]
+  }
+}
+
+interface VertexSettingsGroup {
+  appliesToLabel: string;
+  displayProperty?: string[];
+  color?: string;
+  showLabel?: boolean;
+}
+
+interface GraphSettingsGroup {
+  vertexSettings?: VertexSettingsGroup[];
+}
+
+type GraphViewSettings = GraphSettingsGroup[];
+
+interface PageState {
+  query: string | undefined;
+  results: GraphResults | undefined;
+  errorMessage: string | undefined;
+  view: 'graph' | 'json';
+  isQueryRunning: boolean;
+  runningQueryId: number;
 }
 
 // Messages that are sent from the server to the client